src/test/java/com/ldts/frogger/model/PositionTest.java
package com.ldts.frogger.model;

import net.jqwik.api.ForAll;
import net.jqwik.api.Property;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class PositionTest {
    private Position position;

    @Property
    void getLeft(@ForAll int x, @ForAll int y) {
        assertEquals(x - 1, new Position(x, y).getLeft().x());
        assertEquals(y, new Position(x, y).getLeft().y());
    }

    @Property
    void getRight(@ForAll int x, @ForAll int y) {
        assertEquals(x + 1, new Position(x, y).getRight().x());
        assertEquals(y, new Position(x, y).getRight().y());
    }

    @Property
    void getUp(@ForAll int x, @ForAll int y) {
        assertEquals(x, new Position(x, y).getUp().x());
        assertEquals(y - 1, new Position(x, y).getUp().y());
    }

    @Property
    void getDown(@ForAll int x, @ForAll int y) {
        assertEquals(x, new Position(x, y).getDown().x());
        assertEquals(y + 1, new Position(x, y).getDown().y());
    }
}
